#282 refactor: Render ticker tape only for one depth pages for mobile
