+++ src/components/Header/Header.tsx
@@ -1,5 +1,7 @@
+import isPageDepthOne from "@constants/isPageDepthOne";
 import useResponsiveLayout from "@hooks/useResponsiveLayout";
 import designSystem from "@styles/designSystem";
+import { useLocation } from "react-router-dom";
 import styled from "styled-components";
 import TVTickerTapeWidget from "../TradingViewWidgets/TVTickerTape";
 import HeaderTopD from "./desktop/HeaderTopD";
@@ -7,13 +9,17 @@ import HeaderTopM from "./mobile/HeaderTopM";
 
 export default function Header() {
   const { isDesktop, isMobile } = useResponsiveLayout();
+  const location = useLocation();
 
   return (
     <StyledHeader>
       {isDesktop && <HeaderTopD />}
       {isMobile && <HeaderTopM />}
 
-      <TVTickerTapeWidget />
+      {isDesktop ||
+        (isMobile && isPageDepthOne(location.pathname) && (
+          <TVTickerTapeWidget />
+        ))}
     </StyledHeader>
   );
 }
