+++ src/constants/isPageDepthOne.ts
@@ -1,0 +1,15 @@
+const oneDepthPages = new Set([
+  "/landing",
+  "/dashboard",
+  "/watchlists",
+  "/portfolios",
+  "/portfolio",
+  "/indices",
+  "/stock",
+]);
+
+export default function isPageDepthOne(route: string) {
+  const regex = /\/\w+/;
+  const parsedRoute = route.match(regex);
+  return parsedRoute ? oneDepthPages.has(parsedRoute[0]) : false;
+}
